Add displayio bitmaps to unix build
